refactor(BA-5528): mask inside print_summary, move mask_token to formatter
print_summary previously took a pre-formatted token_display string and
relied on the caller to remember to call mask_token() first. Take the
raw token (or None) instead and mask inside the method — that's the
single place that knows how api_key should appear in the summary.

Move mask_token from utils.py to formatter.py since it's purely
display-side. utils.py is now strictly persistence (save_*,
read_json_file). The remaining mask_token call site in
chat-config set's status output also imports from formatter.

Tests for mask_token move to test_deployment_chat_formatter.py to
match the new home.
